// src/api/embeddedGroupInvite/request/data/invite/invite.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

import { Signer } from './signer';

export interface Invite {
  order: number;
  signers: Signer[];
}