// src/api/embeddedGroupInvite/request/data/invite/document.ts
/*
 * This file is a part of signNow SDK API client.
 *
 * (с) Copyright © 2011-present airSlate Inc. (https://www.signnow.com)
 *
 * For more details on copyright, see LICENSE.md file
 * that was distributed with this source code.
 */

export interface Document {
  id: string;
  action: string;
  role?: string | null;
}